requirements.txt
flask
python-dotenv
uwsgi
requests
python-docx
